fix(MovementUtil): always assuming forward if we're not moving
FINALLY that movement issue should be fixed
these are the type of issues that are a pain yet aren't easy to find
This if statement that we are deleting would assume that if we're moving in **any** direction or strafing in **any** direction, we're going forward. (emphasis on the **any**)
However, new shocking world invention: YOU CAN GO BACKWARDS!!!
So, if you want to go backwards, you go forward when using this.
It's January 14, at 9:56 PM, might as well go to bed lol.
Will push this tomorrow, not sure what the issue ID is for the relevant issue because no internet access.